feat: implement TagBadge component for tag display
